refactor : sorting methods moved to parent class
